cluster: split out feature_update_action
This allows features library to use this type without depending on all
of the cluster library.
