+++ src/v/cluster/types.h
@@ -13,6 +13,7 @@
 
 #include "cluster/cloud_metadata/cluster_manifest.h"
 #include "cluster/errc.h"
+#include "cluster/feature_update_action.h"
 #include "cluster/fwd.h"
 #include "cluster/remote_topic_properties.h"
 #include "cluster/topic_configuration.h"
@@ -1537,37 +1538,6 @@ struct cluster_config_status_cmd_data
     auto serde_fields() { return std::tie(status); }
 
     config_status status;
-};
-
-struct feature_update_action
-  : serde::envelope<
-      feature_update_action,
-      serde::version<0>,
-      serde::compat_version<0>> {
-    static constexpr int8_t current_version = 1;
-    enum class action_t : std::uint16_t {
-        // Notify when a feature is done with preparing phase
-        complete_preparing = 1,
-        // Notify when a feature is made available, either by an administrator
-        // or via auto-activation policy
-        activate = 2,
-        // Notify when a feature is explicitly disabled by an administrator
-        deactivate = 3
-    };
-
-    // Features have an internal bitflag representation, but it is not
-    // meant to be stable for use on the wire, so we refer to features by name
-    ss::sstring feature_name;
-    action_t action;
-
-    friend bool
-    operator==(const feature_update_action&, const feature_update_action&)
-      = default;
-
-    auto serde_fields() { return std::tie(feature_name, action); }
-
-    friend std::ostream&
-    operator<<(std::ostream&, const feature_update_action&);
 };
 
 struct feature_update_cmd_data
