+++ src/v/features/tests/feature_table_test.cc
@@ -10,6 +10,7 @@
  */
 
 #include "base/vlog.h"
+#include "cluster/feature_update_action.h"
 #include "features/feature_table.h"
 #include "features/feature_table_snapshot.h"
 #include "test_utils/fixture.h"
